moving of resolver did to pkg
